Move variable declarations outside of conditionals.
Since all variables are scoped for an entire function, like javascript, whether a conditional hits or not the variable is declared and set to it's default value.
This means that in some places we could see unused variables being declared since that conditional branch never hit.
By declaring them before the conditional and using the same variable in each branch of the conditional we can cut down on the number of variables and thus gas usage.